requirements.txt
ruff
pandas
python-dotenv
influxdb-client
psycopg[binary]